overview: Add shortcuts window
Add shortcuts window since we already have a menu item for it.
There is not many of them currently but we can add them later.

Setting of application-id was moved to construction time to have
resource-base-path set correctly.

The window is inspired by the one from Boxes.

With suggestions from Felipe Borges and Balló György.